+++ prisma/migrations/20240817204837_remove_panel_id_from_cashier/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `panel_id` on the `CASHIERS` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE `CASHIERS` DROP COLUMN `panel_id`;