+++ src/routes/index.ts
@@ -10,6 +10,7 @@ import transactionsRouter from "./transactions";
 import analyticsRouter from "./analytics.router";
 import bonusRouter from "./bonus.router";
 import coinTransferRouter from "./coin-transfer.router";
+import cashierRouter from "./cashier.router";
 
 const mainRouter = Router();
 
@@ -24,5 +25,6 @@ mainRouter.use("/bot", botRouter);
 mainRouter.use("/analytics", analyticsRouter);
 mainRouter.use("/bonus", bonusRouter);
 mainRouter.use("/coin-transfer", coinTransferRouter);
+mainRouter.use("/cashier", cashierRouter);
 
 export default mainRouter;